requirements.txt
appdirs==1.4.4
APScheduler==3.7.0
asgiref==3.3.4
black==21.6b0
click==8.0.1
Django==3.2.4
django-extensions==3.1.3
djangorestframework==3.12.4
mypy-extensions==0.4.3
pathspec==0.8.1
psycopg2
pytz==2021.1
regex==2021.4.4
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
tzlocal==2.1
gunicorn
django-heroku
